% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MEPdata.R
\docType{data}
\name{MEPdata}
\alias{MEPdata}
\title{Information of 45 microexon-tag clusters in plant}
\format{
A list containing three elements:
\itemize{
\item MEPdata$cluster: a data.frame containing microexon-tag cluster 
(cluster), microexon size (zise) and phase (phase), the coding motif (motif),
 number of exons (exons) and the order of microexon (me_order).
\item MEPdata$blocks: an IntegerList containing exon blocks in each cluster.
\item MEPdata$matrix: a list of 45 elements containng the DNA consensus 
matrix of ecach cluster.
}
}
\usage{
data(MEPdata)
}
\description{
MEPdata contains information of 45 microexon-tag clusters in plants.
}
\examples{
data(MEPdata)
names(MEPdata)
sapply(MEPdata, class)
head(MEPdata$cluster)
head(MEPdata$blocks)
names(MEPdata$matrix)

# Microexon cluster 8
MEPdata$cluster[8,]
MEPdata$blocks[[8]]
MEPdata$matrix[[8]]
ggseqlogo::ggseqlogo(MEPdata$matrix[[8]][,53:57])

}
\keyword{datasets}
